safe

+++ form.c
@@ -486,12 +486,14 @@ void delete_after_all(List *list, int index, unsigned char *global_metadata,
     int first_block_num = 0;
     int first_block_start = find_block_start(index, &first_block_num, global_metadata);
 
-    int deleted_blocks = list->count - first_block_num;
-    int delete_start_index = index - first_block_start;
-    int delete_data_size = global_metadata[first_block_num] - delete_start_index;
+    if(strlen(global_metadata) > (first_block_num + 1)){
+        int deleted_blocks = list->count - first_block_num;
+        int delete_start_index = index - first_block_start;
+        int delete_data_size = global_metadata[first_block_num] - delete_start_index;
 
-    removeNodes(list, first_block_num + 1, deleted_blocks);
-    delete_global(global_metadata, first_block_num + 1, deleted_blocks);
+        removeNodes(list, first_block_num + 1, deleted_blocks);
+        delete_global(global_metadata, first_block_num + 1, deleted_blocks);
+    }
 
     deletion_single_block(list, first_block_num, delete_start_index, delete_data_size, global_metadata, enc_key, dec_key);
 }
